Update 1.0.1 Made a fix for terminating buildTools if it was running

+++ src/java/net/lapismc/spigotbuilder/SpigotBuilder.java
@@ -45,7 +45,8 @@ class SpigotBuilder {
         placeEula();
         log("Starting the server");
         startServer(currentSpigotJar);
-        buildTools.destroyForcibly();
+        if (buildTools != null && buildTools.isAlive())
+            buildTools.destroyForcibly();
         System.exit(0);
     }
